backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-google-genai
langchain-community
sqlmodel
asyncpg
python-dotenv
httpx